Move otel4s integration to a separate module
