+++ core/src/main/scala/org/typelevel/keypool/Managed.scala
