+++ core/src/test/scala/org/typelevel/keypool/PoolSpec.scala
